! [Heartbeats] Fixed handling of multiple heartbeats
If a second heartbeat was triggered while we have not received the response to
the first heartbeat, the timeout task of the first heartbeat was overwritten. This
caused the timeout task to be triggered always as there is no way to cancel it
anymore.

We now remember the timeout per heartbeat and support having multiple
pending heartbeats.

Also, there was a race condition when setting up the timeout task. If the
heartbeat returned really fast, it was possible for the timeout task to be
scheduled after the heartbeat response has already been processed.