! [Heartbeats] Fixed multiple heartbeats bug by ensuring only one active heartbeat
If a second heartbeat was triggered while we have not received the response to
the first heartbeat, the timeout task of the first heartbeat was overwritten. This
caused the timeout task to be triggered always as there is no way to cancel it
anymore.

This commit fixes this problem by ensuring only one active heartbeat at any one
time.