! [Operation Timeout] Fixed timeouts not triggering when there is no server response
Duration.between(start, end) measures the duration between the start and the end.
If start is after end, the duration becomes negative, which is never larger than any
configured operation timeout.

Using Instant for timeouts is generally a bad idea because the system clock can
change arbitrarily causing timeouts to be fired prematurly or not at all.
System.nanoTime is guaranteed to be monotonically increasing, making it ideal
for timeout calculations.