Catch exception to keep timer thread running in case of an error
For instance `connectionState()` may throw a NullPointerException in case of a connection being closed (which sets the variable `connection` to `null`) just after the null check happened within `connectionState()`. In this situation no ClientReconnecting / ClientDisconnected event is triggered and any workarounds as described in https://github.com/msemys/esjc/issues/29 fail to reestablish a connection.